requirements.txt
# Core dependencies
numpy
pandas
scikit-learn
matplotlib
seaborn
pyyaml

# Deep learning & NLP
torch
transformers
sentence-transformers
faiss-cpu==1.7.4


# Reinforcement Learning
stable-baselines3
gym

# API
fastapi
uvicorn

# Web UI
streamlit
gradio

# Data processing
tqdm
requests

# Vector Database
pinecone-client==2.2.4


# Testing & Utilities
pytest
black
flake8

python-dotenv

